pyproject.toml
[project]
name = "marine-risk-mapping"
version = "0.1.0"
description = "Marine risk mapping platform combining AIS vessel data, whale sightings, and ocean data to identify at-risk areas"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # ── Core data ────────────────────────────────────
    "numpy>=2.2.0",
    "pandas>=2.3.0",
    "geopandas>=1.1.2",
    "pyarrow>=23.0.1",
    "h3>=4.4.2",
    "scipy>=1.17.1",
    "pyyaml>=6.0",
    # ── Database / IO ────────────────────────────────
    "psycopg2-binary>=2.9.11",
    "duckdb>=1.4.4",
    # ── HTTP / Cloud ─────────────────────────────────
    "httpx>=0.28.1",
    "requests>=2.32.0",
    "boto3>=1.42.52",
    # ── Geospatial ───────────────────────────────────
    "rasterio>=1.5.0",
    "xarray>=2026.2.0",
    "netcdf4>=1.7.4",
    "copernicusmarine>=2.2.3",
    # ── Orchestration ────────────────────────────────
    "dagster>=1.9.0",
    "dagster-webserver>=1.9.0",
    "dagster-dbt>=0.25.0",
    "dbt-postgres>=1.10.0",
    # ── ML / Statistics ──────────────────────────────
    "scikit-learn>=1.8.0",
    "xgboost>=3.2.0",
    "shap>=0.49.1",
    "optuna>=4.7.0",
    "mlflow>=3.10.0",
    # ── Validation / Reporting ───────────────────────
    "pandera>=0.29.0",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "fpdf2>=2.8.6",
    "graphviz>=0.21",
    "pypdf2>=3.0.1",
    # ── Audio classification ─────────────────────────
    "librosa>=0.11.0",
    "soundfile>=0.13.1",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    # ── Data acquisition ─────────────────────────────
    "kaggle>=1.6.0",
    # ── API backend ──────────────────────────────────
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
    "bcrypt>=5.0.0",
    "pyjwt>=2.11.0",
    "email-validator>=2.3.0",
    "slowapi>=0.1.9",
    "openai>=2.26.0",
    "timm>=1.0.25",
    "albumentations>=2.0.8",
]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyterlab>=4.5.4",
    "pre-commit>=4.5.1",
    "pypdf>=6.7.5",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.15.1",
]

[tool.ruff]
target-version = "py312"
line-length = 88
exclude = ["**/*.sh"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "UP",  # pyupgrade (modernise syntax)
    "B",   # flake8-bugbear (common bugs)
    "SIM", # flake8-simplify (simplify code)
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline", "orchestration", "backend"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: requires PostGIS database or external files",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
